Fix more lint issues and tune config
Code fixes:
- Remove unnecessary uint64 conversions
- Add b.Helper() to benchmark helper function

Config tuning:
- Disable ireturn: returning interfaces is intentional
- Disable nilnil: returning nil,nil is correct in some APIs

+++ pkg/extsort/indexbuild.go
@@ -304,7 +304,7 @@ func (b *IndexBuilder) createTierWriter(tierID tiers.ID, _ *PrefixRow) error {
 	b.tierBytesWriters[tierID] = bytesW
 
 	// Backfill zeros for all previous positions
-	for range uint64(b.posCount - 1) {
+	for range b.posCount - 1 {
 		if err := countW.WriteU64(0); err != nil {
 			return fmt.Errorf("backfill tier %s count: %w", info.Name, err)
 		}
